reunite old correlation scripts w/ socorro

+++ scripts/crons/cron_daily_reports.sh
@@ -1,40 +1,34 @@
-#!/bin/bash
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#! /bin/sh
 
+[[ -f /tmp/daily_urls.lock ]] && exit 1
+
+touch /tmp/daily_urls.lock
 
 . /etc/socorro/socorrorc
 
-SCRIPT_RUN_DATE=`date -d "1 days ago" '+%Y-%m-%d'`
+REPORT_DATE="1 days ago"
+
 if [ -n "$1" ]
 then
-  SCRIPT_RUN_DATE=$1
+  REPORT_DATE=$1
 fi
 
-NAME=`basename $0 .sh`
-lock $NAME
-$PYTHON ${APPDIR}/scripts/startDailyUrl.py --day=${SCRIPT_RUN_DATE}
-fatal $? "Could not run startDailyUrl"
+SCRIPT_RUN_DATE=`date -d "$REPORT_DATE" '+%Y-%m-%d'`
+$PYTHON /data/socorro/application/scripts/startDailyUrl.py --day=$SCRIPT_RUN_DATE
 
-PUB_DATA_FILE=`date -d "1 days ago" '+%Y%m%d-pub-crashdata.csv.gz'`
-SCRIPT_RUN_DATE=`date -d "1 days ago" '+%Y%m%d'`
-scp ${PUB_DATA_FILE} ${PUB_USER}@${PUB_SERVER}:${PUB_LOCATION}/${SCRIPT_RUN_DATE}/
-EXIT_CODE=$((EXIT_CODE+$?))
+#DATA_FILE=`date -d "$REPORT_DATE" '+%Y%m%d-crashdata.csv.gz'`
+#scp ${HOME}/${DATA_FILE} bacula@10.22.72.131:/data/security_group/crash_urls/
+#scp ${HOME}/${DATA_FILE} mozauto@sisyphus.bughunter.ateam.phx1.mozilla.com:/work/mozilla/crash-reports/
+#ssh bacula@10.22.72.131 'chmod 640 /data/security_group/crash_urls/*'
+#ssh mozauto@sisyphus.bughunter.ateam.phx1.mozilla.com 'chmod 640 /work/mozilla/crash-reports/*'
+#mv ${HOME}/${DATA_FILE} /tmp
+
+DATA_FILE=`date -d "$REPORT_DATE" '+%Y%m%d-pub-crashdata.csv.gz'`
+SCRIPT_RUN_DATE=`date -d "$REPORT_DATE" '+%Y%m%d'`
+#scp $DATA_FILE bacula@people.mozilla.org:/var/www/html/crash_analysis/$SCRIPT_RUN_DATE/
+
+mkdir -p /mnt/crashanalysis/crash_analysis/$SCRIPT_RUN_DATE/
+cp $DATA_FILE /mnt/crashanalysis/crash_analysis/$SCRIPT_RUN_DATE/
 mv ${HOME}/${DATA_FILE} /tmp
 
-PRIV_DATA_FILE=`date -d "1 days ago" '+%Y%m%d-crashdata.csv.gz'`
-scp ${PRIV_DATA_FILE} ${PRIV_USER}@${PRIV_SERVER}:${PRIV_LOCATION}/
-EXIT_CODE=$((EXIT_CODE+$?))
-ssh ${PRIV_USER}@${PRIV_SERVER} "chmod 640 ${PRIV_LOCATION}/*"
-EXIT_CODE=$((EXIT_CODE+$?))
-mv ${HOME}/${DATA_FILE} /tmp
-
-unlock $NAME
-
-if [ "$EXIT_CODE" != 0 ]
-then
-  echo "ERROR: problems copying files"
-  exit 1
-fi
-
+rm -f /tmp/daily_urls.lock
